requirements.txt
--constraint "https://raw.githubusercontent.com/apache/airflow/constraints-2.7.2/constraints-3.11.txt"
apache-airflow==2.7.2
s3fs
boto3
pandas
requests